requirements.txt
# Core Django Framework
Django>=5.0,<6.0
python-dotenv>=1.0.1

# REST API Framework (Task 1.1, 1.2, 1.3, 1.4)
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.5,<6.0
django-filter>=25.0
drf-yasg>=1.21,<2.0

# CORS Support (Task 1.6)
django-cors-headers>=4.9,<5.0

# Async Tasks and Background Jobs (Task 1.5)
celery>=5.6,<6.0
redis>=7.1,<8.0
django-celery-beat>=2.8,<3.0

# Database
psycopg2-binary>=2.9

# Static Files and Media
whitenoise>=6.6.0
Pillow>=10.2.0

# Frontend and UI
django-jazzmin>=3.0
django-browser-reload>=1.21
django-htmx>=1.17.0
django-widget-tweaks>=1.5.0
django-compressor>=4.4
django-tailwind>=3.8.0

# Production Server
gunicorn>=21.2.0
